Check idFileOnline is missing
https://bugs.launchpad.net/subdownloader/+bug/722084 comment 8

Apparently some files are missing the idFileOnline, and they crash the party when they are added to the download matrix in filter.py (I've only encountered this on files that shouldn't be downloaded in any case since they have local subs, might be a problem if --sol is specified). I made a simple check for the existance of this field before adding the file and that resolves the crashes.